# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/role": {
      "post": {
        "operationId": "RoleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "get": {
        "operationId": "RoleController_findAll",
        "parameters": [
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "perPage",
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "sortBy",
            "in": "query",
            "schema": {
              "default": "id.ASC",
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/role/{id}": {
      "get": {
        "operationId": "RoleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "patch": {
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "RoleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
    "/user": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "perPage",
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "sortBy",
            "in": "query",
            "schema": {
              "default": "id.ASC",
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
    "/warehouse/type": {
      "post": {
        "operationId": "WarehouseController_createType",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWarehouseTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Warehouse"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "get": {
        "operationId": "WarehouseController_findAllType",
        "parameters": [
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "perPage",
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "sortBy",
            "in": "query",
            "schema": {
              "default": "id.ASC",
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Warehouse"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/warehouse": {
      "post": {
        "operationId": "WarehouseController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWarehouseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Warehouse"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "get": {
        "operationId": "WarehouseController_findAll",
        "parameters": [
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "perPage",
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "sortBy",
            "in": "query",
            "schema": {
              "default": "id.ASC",
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Warehouse"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/warehouse/type/{id}": {
      "get": {
        "operationId": "WarehouseController_findOneType",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Warehouse"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "patch": {
        "operationId": "WarehouseController_updateType",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Warehouse"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "WarehouseController_removeType",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Warehouse"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/warehouse/{id}": {
      "get": {
        "operationId": "WarehouseController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Warehouse"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "patch": {
        "operationId": "WarehouseController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Warehouse"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "WarehouseController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Warehouse"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Warehouse MS Swagger",
    "description": "The Warehouse MS API documents",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "authorization": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    },
    "schemas": {
...
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "name",
          "description",
          "permissionIds"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "roleId": {
            "type": "number"
          },
          "departmentId": {
            "type": "number"
          },
          "fullName": {
            "type": "string"
          },
          "areaCode": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "email",
          "roleId",
          "departmentId",
          "fullName",
          "areaCode",
          "phone",
          "address",
          "birthday",
          "gender"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "roleId": {
            "type": "number"
          },
          "departmentId": {
            "type": "number"
          },
          "fullName": {
            "type": "string"
          },
          "areaCode": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          }
        }
      },
...
      "CreateWarehouseTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "CreateWarehouseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "typeId": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "code",
          "typeId",
          "description",
          "address"
        ]
      },
      "UpdateWarehouseTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "UpdateWarehouseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "typeId": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
...
    }
  }
}